refactor: move DispatcherServlet
